Finding min and max in an array
Write a program which takes as input a sequence  of integers, and outputs the smallest and largest elements of the sequence. Your program must not use the built-in functions available to compute min and max, and must compute the output in linear time.

Input Format

The input consists of two lines. The first line of the input is the number of integers  in the input array. The second line of the input is a sequence  of  integers. Consecutive elements of the sequence are separated by a space.